Provider and components
Added a main component App.js file.
Linked App.js to the Sass/Scss style file.

Added the Provider to js/index.js file from react-redux.

Added a Provider tag to wrap around the App component in the ReactDOM.render so that the store is now linked to all the components