Passing Data to Component
Store is accessible so going to render the data in a component.

In containers/user-list.js in the parent component we are going to render we made a child component called createListItems.

Added createListItems functional component which will loop through the state with a map call and give us a list of the first and last name of each item in the state.